# js/apps/devtools-app/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KotoneBot DevTools",
    "version": "0.1.0"
  },
  "paths": {
    "/api/fs/list_dir": {
      "get": {
        "tags": [
          "filesystem"
        ],
        "summary": "List Dir",
        "operationId": "list_dir_api_fs_list_dir_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ListDirResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fs/read_text": {
      "get": {
        "tags": [
          "filesystem"
        ],
        "summary": "Read Text",
        "operationId": "read_text_api_fs_read_text_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ReadTextResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fs/write_text": {
      "put": {
        "tags": [
          "filesystem"
        ],
        "summary": "Write Text",
        "operationId": "write_text_api_fs_write_text_put",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WriteTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fs/rename": {
      "post": {
        "tags": [
          "filesystem"
        ],
        "summary": "Rename Path",
        "operationId": "rename_path_api_fs_rename_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenamePathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RenameResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fs/copy_file": {
      "post": {
        "tags": [
          "filesystem"
        ],
        "summary": "Copy File",
        "operationId": "copy_file_api_fs_copy_file_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopyFileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_CopyResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fs/upload_file": {
      "post": {
        "tags": [
          "filesystem"
        ],
        "summary": "Upload File",
        "operationId": "upload_file_api_fs_upload_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_fs_upload_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UploadResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fs/reveal_in_explorer": {
      "post": {
        "tags": [
          "filesystem"
        ],
        "summary": "Reveal In Explorer",
        "operationId": "reveal_in_explorer_api_fs_reveal_in_explorer_post",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_NoneType_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meta/index/update": {
      "post": {
        "tags": [
          "meta"
        ],
        "summary": "Update Meta Index",
        "operationId": "update_meta_index_api_meta_index_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_SymbolUpdateResultModel_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/meta/variant/clone_to_image": {
      "post": {
        "tags": [
          "variant"
        ],
        "summary": "Clone Variant To Image",
        "operationId": "clone_variant_to_image_api_meta_variant_clone_to_image_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneVariantToImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_VariantCloneResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/variant/import/precheck_path": {
      "post": {
        "tags": [
          "variant"
        ],
        "summary": "Precheck Variant Import Path",
        "operationId": "precheck_variant_import_path_api_meta_variant_import_precheck_path_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_precheck_variant_import_path_api_meta_variant_import_precheck_path_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_VariantImportPrecheckResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/variant/import_image": {
      "post": {
        "tags": [
          "variant"
        ],
        "summary": "Import Variant Image",
        "operationId": "import_variant_image_api_meta_variant_import_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_variant_image_api_meta_variant_import_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_VariantImportResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/variant/copy_selected_prefab/precheck": {
      "post": {
        "tags": [
          "variant"
        ],
        "summary": "Precheck Copy Selected Prefab To Variant",
        "operationId": "precheck_copy_selected_prefab_to_variant_api_meta_variant_copy_selected_prefab_precheck_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrecheckCopySelectedPrefabToVariantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_CopyPrefabPrecheckResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/meta/variant/copy_selected_prefab": {
      "post": {
        "tags": [
          "variant"
        ],
        "summary": "Copy Selected Prefab To Variant",
        "operationId": "copy_selected_prefab_to_variant_api_meta_variant_copy_selected_prefab_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CopySelectedPrefabToVariantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_CopyPrefabResult_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_variant_image_api_meta_variant_import_image_post": {
        "properties": {
          "baseImagePath": {
            "type": "string",
            "title": "Baseimagepath"
          },
          "variant": {
            "type": "string",
            "title": "Variant"
          },
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          },
          "deleteExistingTarget": {
            "type": "boolean",
            "title": "Deleteexistingtarget",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "baseImagePath",
          "variant",
          "image"
        ],
        "title": "Body_import_variant_image_api_meta_variant_import_image_post"
      },
...
      "Body_precheck_variant_import_path_api_meta_variant_import_precheck_path_post": {
        "properties": {
          "sourceMetaPath": {
            "type": "string",
            "title": "Sourcemetapath"
          },
          "baseImagePath": {
            "type": "string",
            "title": "Baseimagepath"
          },
          "variant": {
            "type": "string",
            "title": "Variant"
          },
          "image": {
            "type": "string",
            "format": "binary",
            "title": "Image"
          }
        },
        "type": "object",
        "required": [
          "sourceMetaPath",
          "baseImagePath",
          "variant",
          "image"
        ],
        "title": "Body_precheck_variant_import_path_api_meta_variant_import_precheck_path_post"
      },
...
      "Body_upload_file_api_fs_upload_file_post": {
        "properties": {
          "targetPath": {
            "type": "string",
            "title": "Targetpath"
          },
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "targetPath",
          "file"
        ],
        "title": "Body_upload_file_api_fs_upload_file_post"
      },
...
      "CloneVariantToImageRequest": {
        "properties": {
          "sourceMetaPath": {
            "type": "string",
            "title": "Sourcemetapath"
          },
          "targetImagePath": {
            "type": "string",
            "title": "Targetimagepath"
          },
          "variant": {
            "type": "string",
            "title": "Variant"
          },
          "forceOverwrite": {
            "type": "boolean",
            "title": "Forceoverwrite",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "sourceMetaPath",
          "targetImagePath",
          "variant"
        ],
        "title": "CloneVariantToImageRequest"
      },
...
      "CopyFileRequest": {
        "properties": {
          "sourcePath": {
            "type": "string",
            "title": "Sourcepath"
          },
          "targetPath": {
            "type": "string",
            "title": "Targetpath"
          }
        },
        "type": "object",
        "required": [
          "sourcePath",
          "targetPath"
        ],
        "title": "CopyFileRequest",
        "description": "文件拷贝覆盖请求。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RenamePathRequest": {
        "properties": {
          "sourcePath": {
            "type": "string",
            "title": "Sourcepath"
          },
          "targetPath": {
            "type": "string",
            "title": "Targetpath"
          }
        },
        "type": "object",
        "required": [
          "sourcePath",
          "targetPath"
        ],
        "title": "RenamePathRequest",
        "description": "单文件重命名请求。"
      },
...
      "ResponseModel_CopyResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CopyResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[CopyResult]"
      },
...
      "ResponseModel_ListDirResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ListDirResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[ListDirResponse]"
      },
...
      "ResponseModel_NoneType_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "type": "null",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[NoneType]"
      },
...
      "ResponseModel_ReadTextResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadTextResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[ReadTextResponse]"
      },
...
      "ResponseModel_RenameResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RenameResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[RenameResult]"
      },
...
      "ResponseModel_SymbolUpdateResultModel_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SymbolUpdateResultModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[SymbolUpdateResultModel]"
      },
...
      "ResponseModel_UploadResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[UploadResult]"
      },
      "ResponseModel_VariantCloneResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VariantCloneResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[VariantCloneResult]"
      },
      "ResponseModel_VariantImportPrecheckResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VariantImportPrecheckResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[VariantImportPrecheckResult]"
      },
      "ResponseModel_VariantImportResult_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VariantImportResult"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ResponseModel[VariantImportResult]"
      },
...
      "UpdateIndexRequest": {
        "properties": {
          "metaPath": {
            "type": "string",
            "title": "Metapath"
          }
        },
        "type": "object",
        "required": [
          "metaPath"
        ],
        "title": "UpdateIndexRequest"
      },
...
      "WriteTextRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "WriteTextRequest"
      }
    }
  }
}